Import dotenv/config in webpack config
Load environment variables from env file before webpack imports the
constants module. The constants module will throw if required
environment variables are unset.

This fixes the webpack build, which was broken due to the introduction
of `just` in the constants module to appease the flow
typechecker.

Builds will still fail if required environment variables are not found
in either the process environment or the env file. This is intentional
and ensures that incomplete configurations cannot be deployed.

Note that webpack, via the Dotenv plugin, also performs text
replacement in the bundle for occurrences of environment
variables.

The present change does not affect this text replacement. However,
since the env file has already been loaded into the process
environment, the Dotenv plugin will now effectively only use the
process environment.