Switch to axios
This article describes why it is recommended to use an HTTP toolkit
instead of the Fetch API in application code.

  https://medium.com/@shahata/why-i-wont-be-using-fetch-api-in-my-apps-6900e6c6fe78

Also, we could not get our tests to work properly with fetch and
React. This tutorial on fetching data in React components uses axios,
and provides working examples for testing.

  https://www.robinwieruch.de/react-fetching-data/

The tutorial uses sinon for mocking. We translate the mocking code to
Jest using the following resource:

  https://github.com/maurocarrero/sinon-jest-cheatsheet

Closes: #15