Add snapshot tests for App
This commit adds snapshot tests for both routes in the App, using
`MemoryRouter` to invoke specific routes via the `initialEntries`
prop.

We provide an `AppWithoutRouter` component, which is identical to
`App` except that it is not wrapped by `BrowserRouter`. This is
required because the component cannot be wrapped by both
`BrowserRouter` and `MemoryRouter` at the same time.

Snapshot tests fail when `initialEntries` contains only an array of
pathnames. This is due to randomly generated keys being used at each
test invocation. To avoid this, we provide a fixed `key` to
`MemoryRouter` entries. The technique is described here:

  https://github.com/ReactTraining/react-router/issues/5579#issuecomment-333401692

An alternative technique sets the `keyLength` attribute of router
entries to zero.

The snapshot test for the default route uses `App` instead of
`AppWithoutRouter` in order to provide full test coverage for the
module.