Do not compile files in __tests__ directories
Update the babel --ignore option to match __tests__ directories,
rather than files ending in .test.js. The .test.js extension is no
longer used (see 750521a).

Note that this configuration option must be passed on the command
line. It cannot be in .babelrc, because jest uses babel to load the
test modules.