Add INTERNAL_API_URL environment variable
If INTERNAL_API_URL is set, requests of the server to the API will use it
instead of API_URL. This setting can be used to allow requests from the server
to the API to use the internal network when both are deployed in the same
environment.

This setting also allows Docker setups where the browser cannot use the API
container name. The container name is only resolvable on the Docker bridge
network. In this scenario, the server can use the container name, but the client
uses a port on localhost, where the API must be exposed explicitly.

This commit also adds a constant CLIENT_API_URL, containing the API URL for the
client. This constant is used only to set `window.API_URL` for the client.