[SignIn] Fix style of sign up link
After upgrading to Material UI 4.x, the specified style was no longer applied to
the link:

- The link was not centered
- The top margin was not applied to the link

Fix this issue by using the new Material UI Link component, and by formatting
the link as a block element. Note that this also displays the link blue and
without decoration, rather than black and underlined.